Revert "[Epmgcipgab 66]"
