--- messages/en.json
@@ -1,6 +1,0 @@
-{
-  "Index": {
-    "title": "Home",
-    "description": "This is the home page."
-  }
-}
