--- messages/ru.json
@@ -1,6 +1,0 @@
-{
-  "Index": {
-    "title": "[ru] Home",
-    "description": "[ru] This is the home page."
-  }
-}
