+++ next.config.mjs
@@ -1,8 +1,4 @@
-import createNextIntlPlugin from "next-intl/plugin";
-
-const withNextIntl = createNextIntlPlugin();
-
-/** @type {import('next/dist/server/config').NextConfig} */
+/** @type {import('next').NextConfig} */
 const nextConfig = {};
 
-export default withNextIntl(nextConfig);
+export default nextConfig;
