+++ package.json
@@ -11,7 +11,6 @@
   "dependencies": {
     "firebase": "^10.11.1",
     "next": "14.2.3",
-    "next-intl": "^3.12.2",
     "react": "^18",
     "react-dom": "^18"
   },
