+++ src/app/page.tsx
@@ -1,0 +1,7 @@
+export default function Home() {
+  return (
+    <main>
+      <p>uz-gabt-webapp !</p>
+    </main>
+  );
+}
