--- src/middleware.ts
@@ -1,25 +1,0 @@
-
-import createMiddleware from 'next-intl/middleware';
-import {pathnames, locales, localePrefix, defaultLocale} from './config';
-
-export default createMiddleware({
-  defaultLocale,
-  locales,
-  pathnames,
-  localePrefix
-});
-
-export const config = {
-  matcher: [
-    // Enable a redirect to a matching locale at the root
-    '/',
-
-    // Set a cookie to remember the previous locale for
-    // all requests that have a locale prefix
-    '/(ru|en)/:path*',
-
-    // Enable redirects that add missing locales
-    // (e.g. `/pathnames` -> `/en/pathnames`)
-    '/((?!_next|_vercel|.*\\..*).*)'
-  ]
-};
